# make Printf's general format C11 compliant