--- a/stdlib/Printf/src/Printf.jl
+++ b/stdlib/Printf/src/Printf.jl
@@ -421,9 +421,28 @@ const __BIG_FLOAT_MAX__ = 8192
     elseif T == Val{'f'} || T == Val{'F'}
         newpos = Ryu.writefixed(buf, pos, x, prec, plus, space, hash, UInt8('.'))
     elseif T == Val{'g'} || T == Val{'G'}
+        # C11-compliant general format
         prec = prec == 0 ? 1 : prec
-        x = round(x, sigdigits=prec)
-        newpos = Ryu.writeshortest(buf, pos, x, plus, space, hash, prec, T == Val{'g'} ? UInt8('e') : UInt8('E'), true, UInt8('.'))
+        # format the value in scientific notation and parse the exponent part
+        exp = let p = Ryu.writeexp(buf, pos, x, prec)
+            b1, b2, b3, b4 = buf[p-4], buf[p-3], buf[p-2], buf[p-1]
+            Z = UInt8('0')
+            if b1 == UInt8('e')
+                # two-digit exponent
+                sign = b2 == UInt8('+') ? 1 : -1
+                exp = 10 * (b3 - Z) + (b4 - Z)
+            else
+                # three-digit exponent
+                sign = b1 == UInt8('+') ? 1 : -1
+                exp = 100 * (b2 - Z) + 10 * (b3 - Z) + (b4 - Z)
+            end
+            flipsign(exp, sign)
+        end
+        if -4 ≤ exp < prec
+            newpos = Ryu.writefixed(buf, pos, x, prec - (exp + 1), plus, space, hash, UInt8('.'), !hash)
+        else
+            newpos = Ryu.writeexp(buf, pos, x, prec - 1, plus, space, hash, T == Val{'g'} ? UInt8('e') : UInt8('E'), UInt8('.'), !hash)
+        end
     elseif T == Val{'a'} || T == Val{'A'}
         x, neg = x < 0 || x === -Base.zero(x) ? (-x, true) : (x, false)
         newpos = pos