In general, rounding a floating-point number before formatting is considered harmful, because it may introduce some rounding error. As an extreme example, the largest normal number `1.7976931348623157e308` cannot be correctly formatted with precision 1 if we round it before sending it to Ryu, because the rounded result would be larger than the largest representable number, as shown below:
```
julia> using Printf

julia> @sprintf "%.1g" floatmax()  # expected: "2e+308"
"0"
```

The [N1570](http://www.open-std.org/jtc1/sc22/wg14/www/docs/n1570.pdf) (the final draft of C11, §7.21.6.1, page 313) defines the general format (`%g` and `%G`) as follows:
![image](https://user-images.githubusercontent.com/905683/116846497-5aac3480-ac23-11eb-8085-842b0dab11dd.png)

This PR makes the general format follow this standard. It may make the format slower because the current algorithm formats the same number twice in order to get the exponent part. I couldn't come up with a better way.

Also, this PR fixes #39748.