@quinnj 2021-05-03
Hmmmm, would you mind posting some benchmarks? It seems unfortunate that we'd have to call `Ryu.writexp` twice in certain cases, and `Ryu.writeexp` then `Ryu.writefixed` in others. Admittedly I've been slowing in responding to the original issue here, but I was imagining doing the work inside `Ryu.writeshortest` itself, where there are already some rounding routines and such.