@bicycle1885 2021-05-03
I quickly benchmarked the performance with the following script:
```julia
using Random
Random.seed!(1234)
N = 500_000
x = [randn(N); exp.(randn(N) * 100)]
shuffle!(x)

using Printf
function g6(out, x)
    for xx in x
        @printf out "%.6g" xx
    end
end

using BenchmarkTools
@btime g6($devnull, $x)
```

PR:
```
~/w/julia (printf-general|…) $ ./julia g.jl
  223.687 ms (2000000 allocations: 411.99 MiB)
~/w/julia (printf-general|…) $ ./julia g.jl
  227.702 ms (2000000 allocations: 411.99 MiB)
~/w/julia (printf-general|…) $ ./julia g.jl
  226.366 ms (2000000 allocations: 411.99 MiB)
```

master (8fdd1371f8):
```
~/w/julia (master|…) $ ./julia g.jl
  216.545 ms (2000000 allocations: 411.99 MiB)
~/w/julia (master|…) $ ./julia g.jl
  215.115 ms (2000000 allocations: 411.99 MiB)
~/w/julia (master|…) $ ./julia g.jl
  214.878 ms (2000000 allocations: 411.99 MiB)
```

Surely it slowed the throughput down, but the difference is not devastating, which was much smaller than I expected. I agree that this should be fixed inside Ryu, but if it takes time I think this slowdown is tolerable for the time being.