@bicycle1885 2021-05-03
Even if fixed-point and scientific cases are segregated, the result seems to be the same:
```julia
using Random
Random.seed!(1234)
N = 500_000
x = randn(N)
y = exp.(randn(N) * 100)
z = shuffle!([x; y])

using Printf
function g6(out, x)
    for xx in x
        @printf out "%.6g" xx
    end
end

using BenchmarkTools
@btime g6($devnull, $x)
@btime g6($devnull, $y)
@btime g6($devnull, $z)
```

PR:
```
~/w/julia (printf-general|…) $ ./julia g.jl
  100.871 ms (1000000 allocations: 205.99 MiB)
  123.202 ms (1000000 allocations: 205.99 MiB)
  229.081 ms (2000000 allocations: 411.99 MiB)
~/w/julia (printf-general|…) $ ./julia g.jl
  98.763 ms (1000000 allocations: 205.99 MiB)
  120.594 ms (1000000 allocations: 205.99 MiB)
  225.540 ms (2000000 allocations: 411.99 MiB)
```

master:
```
~/w/julia (master|…) $ ./julia g.jl
  108.733 ms (1000000 allocations: 205.99 MiB)
  105.132 ms (1000000 allocations: 205.99 MiB)
  218.649 ms (2000000 allocations: 411.99 MiB)
~/w/julia (master|…) $ ./julia g.jl
  106.367 ms (1000000 allocations: 205.99 MiB)
  104.395 ms (1000000 allocations: 205.99 MiB)
  215.350 ms (2000000 allocations: 411.99 MiB)
```