@quinnj 2021-05-11
Here are the results I get.

master:
```julia
julia> @btime g6($devnull, $x)
  163.121 ms (2000000 allocations: 411.99 MiB)

  julia> @btime g6($devnull, $x)
  80.925 ms (1000000 allocations: 205.99 MiB)

julia> @btime g6($devnull, $y)
  78.093 ms (1000000 allocations: 205.99 MiB)

julia> @btime g6($devnull, $z)
  165.143 ms (2000000 allocations: 411.99 MiB)
```

PR:
```julia
 julia> @btime g6($devnull, $x)
  194.712 ms (2000000 allocations: 411.99 MiB)

julia> @btime g6($devnull, $x)
  81.821 ms (1000000 allocations: 205.99 MiB)

julia> @btime g6($devnull, $y)
  107.219 ms (1000000 allocations: 205.99 MiB)

julia> @btime g6($devnull, $z)
  196.549 ms (2000000 allocations: 411.99 MiB)
```