@bicycle1885 2021-05-11
Yes, I think this should be backported. `%g` of the latest release is unreliable and I'm not sure when it fails. If this is going to be backported, #40685 and #40755 should be included as well (so I've added backport labels to them).